+++ packages/signpdf/src/signpdf.js
@@ -1,9 +1,9 @@
-import forge from 'node-forge';
 import {
     DEFAULT_BYTE_RANGE_PLACEHOLDER,
     removeTrailingNewLine,
     findByteRange,
     SignPdfError,
+    Signer,
 } from '@signpdf/utils';
 
 /**
@@ -20,30 +20,23 @@ export class SignPdf {
 
     /**
      * @param {Buffer} pdfBuffer
-     * @param {Buffer} p12Buffer
+     * @param {Signer} signer
      * @param {SignerOptions} additionalOptions
-     * @returns {Buffer}
+     * @returns {Promise<Buffer>}
      */
-    sign(
+    async sign(
         pdfBuffer,
-        p12Buffer,
-        additionalOptions = {},
+        signer,
     ) {
-        const options = {
-            asn1StrictParsing: false,
-            passphrase: '',
-            ...additionalOptions,
-        };
-
         if (!(pdfBuffer instanceof Buffer)) {
             throw new SignPdfError(
                 'PDF expected as Buffer.',
                 SignPdfError.TYPE_INPUT,
             );
         }
-        if (!(p12Buffer instanceof Buffer)) {
+        if (!(signer instanceof Signer)) {
             throw new SignPdfError(
-                'p12 certificate expected as Buffer.',
+                'Signer implementation expected.',
                 SignPdfError.TYPE_INPUT,
             );
         }
@@ -89,84 +82,9 @@ export class SignPdf {
             pdf.slice(byteRange[2], byteRange[2] + byteRange[3]),
         ]);
 
-        // Convert Buffer P12 to a forge implementation.
-        const forgeCert = forge.util.createBuffer(p12Buffer.toString('binary'));
-        const p12Asn1 = forge.asn1.fromDer(forgeCert);
-        const p12 = forge.pkcs12.pkcs12FromAsn1(
-            p12Asn1,
-            options.asn1StrictParsing,
-            options.passphrase,
-        );
-
-        // Extract safe bags by type.
-        // We will need all the certificates and the private key.
-        const certBags = p12.getBags({
-            bagType: forge.pki.oids.certBag,
-        })[forge.pki.oids.certBag];
-        const keyBags = p12.getBags({
-            bagType: forge.pki.oids.pkcs8ShroudedKeyBag,
-        })[forge.pki.oids.pkcs8ShroudedKeyBag];
-
-        const privateKey = keyBags[0].key;
-        // Here comes the actual PKCS#7 signing.
-        const p7 = forge.pkcs7.createSignedData();
-        // Start off by setting the content.
-        p7.content = forge.util.createBuffer(pdf.toString('binary'));
-
-        // Then add all the certificates (-cacerts & -clcerts)
-        // Keep track of the last found client certificate.
-        // This will be the public key that will be bundled in the signature.
-        let certificate;
-        Object.keys(certBags).forEach((i) => {
-            const {publicKey} = certBags[i].cert;
-
-            p7.addCertificate(certBags[i].cert);
-
-            // Try to find the certificate that matches the private key.
-            if (privateKey.n.compareTo(publicKey.n) === 0
-                && privateKey.e.compareTo(publicKey.e) === 0
-            ) {
-                certificate = certBags[i].cert;
-            }
-        });
-
-        if (typeof certificate === 'undefined') {
-            throw new SignPdfError(
-                'Failed to find a certificate that matches the private key.',
-                SignPdfError.TYPE_INPUT,
-            );
-        }
-
-        // Add a sha256 signer. That's what Adobe.PPKLite adbe.pkcs7.detached expects.
-        // Note that the authenticatedAttributes order is relevant for correct
-        // EU signature validation:
-        // https://ec.europa.eu/digital-building-blocks/DSS/webapp-demo/validation
-        p7.addSigner({
-            key: privateKey,
-            certificate,
-            digestAlgorithm: forge.pki.oids.sha256,
-            authenticatedAttributes: [
-                {
-                    type: forge.pki.oids.contentType,
-                    value: forge.pki.oids.data,
-                }, {
-                    type: forge.pki.oids.signingTime,
-                    // value can also be auto-populated at signing time
-                    // We may also support passing this as an option to sign().
-                    // Would be useful to match the creation time of the document for example.
-                    value: new Date(),
-                }, {
-                    type: forge.pki.oids.messageDigest,
-                    // value will be auto-populated at signing time
-                },
-            ],
-        });
-
-        // Sign in detached mode.
-        p7.sign({detached: true});
+        const raw = await signer.sign(pdf);
 
         // Check if the PDF has a good enough placeholder to fit the signature.
-        const raw = forge.asn1.toDer(p7.toAsn1()).getBytes();
         // placeholderLength represents the length of the HEXified symbols but we're
         // checking the actual lengths.
         if ((raw.length * 2) > placeholderLength) {