+++ packages/utils/src/index.js
@@ -3,3 +3,4 @@ export * from './extractSignature';
 export * from './findByteRange';
 export * from './removeTrailingNewLine';
 export * from './SignPdfError';
+export * from './Signer';