+++ packages/internal-utils/index.js
@@ -1,0 +1,7 @@
+var readTestResource = require('./readTestResource');
+var createPdfkitDocument = require('./createPdfkitDocument');
+
+module.exports = {
+    readTestResource: readTestResource,  
+    createPdfkitDocument: createPdfkitDocument,  
+}