+++ packages/signpdf/src/signpdf.js
@@ -1,0 +1,118 @@
+import {
+    DEFAULT_BYTE_RANGE_PLACEHOLDER,
+    removeTrailingNewLine,
+    findByteRange,
+    SignPdfError,
+    Signer,
+} from '@signpdf/utils';
+
+/**
+ * @typedef {object} SignerOptions
+ * @prop {string} [passphrase]
+ * @prop {boolean} [asn1StrictParsing]
+ */
+
+export class SignPdf {
+    constructor() {
+        this.byteRangePlaceholder = DEFAULT_BYTE_RANGE_PLACEHOLDER;
+        this.lastSignature = null;
+    }
+
+    /**
+     * @param {Buffer} pdfBuffer
+     * @param {Signer} signer
+     * @param {SignerOptions} additionalOptions
+     * @returns {Promise<Buffer>}
+     */
+    async sign(
+        pdfBuffer,
+        signer,
+    ) {
+        if (!(pdfBuffer instanceof Buffer)) {
+            throw new SignPdfError(
+                'PDF expected as Buffer.',
+                SignPdfError.TYPE_INPUT,
+            );
+        }
+        if (!(signer instanceof Signer)) {
+            throw new SignPdfError(
+                'Signer implementation expected.',
+                SignPdfError.TYPE_INPUT,
+            );
+        }
+
+        let pdf = removeTrailingNewLine(pdfBuffer);
+
+        // Find the ByteRange placeholder.
+        const {byteRangePlaceholder} = findByteRange(pdf);
+
+        if (!byteRangePlaceholder) {
+            throw new SignPdfError(
+                `Could not find empty ByteRange placeholder: ${byteRangePlaceholder}`,
+                SignPdfError.TYPE_PARSE,
+            );
+        }
+
+        const byteRangePos = pdf.indexOf(byteRangePlaceholder);
+
+        // Calculate the actual ByteRange that needs to replace the placeholder.
+        const byteRangeEnd = byteRangePos + byteRangePlaceholder.length;
+        const contentsTagPos = pdf.indexOf('/Contents ', byteRangeEnd);
+        const placeholderPos = pdf.indexOf('<', contentsTagPos);
+        const placeholderEnd = pdf.indexOf('>', placeholderPos);
+        const placeholderLengthWithBrackets = (placeholderEnd + 1) - placeholderPos;
+        const placeholderLength = placeholderLengthWithBrackets - 2;
+        const byteRange = [0, 0, 0, 0];
+        byteRange[1] = placeholderPos;
+        byteRange[2] = byteRange[1] + placeholderLengthWithBrackets;
+        byteRange[3] = pdf.length - byteRange[2];
+        let actualByteRange = `/ByteRange [${byteRange.join(' ')}]`;
+        actualByteRange += ' '.repeat(byteRangePlaceholder.length - actualByteRange.length);
+
+        // Replace the /ByteRange placeholder with the actual ByteRange
+        pdf = Buffer.concat([
+            pdf.slice(0, byteRangePos),
+            Buffer.from(actualByteRange),
+            pdf.slice(byteRangeEnd),
+        ]);
+
+        // Remove the placeholder signature
+        pdf = Buffer.concat([
+            pdf.slice(0, byteRange[1]),
+            pdf.slice(byteRange[2], byteRange[2] + byteRange[3]),
+        ]);
+
+        const raw = await signer.sign(pdf);
+
+        // Check if the PDF has a good enough placeholder to fit the signature.
+        // placeholderLength represents the length of the HEXified symbols but we're
+        // checking the actual lengths.
+        if ((raw.length * 2) > placeholderLength) {
+            throw new SignPdfError(
+                `Signature exceeds placeholder length: ${raw.length * 2} > ${placeholderLength}`,
+                SignPdfError.TYPE_INPUT,
+            );
+        }
+
+        let signature = Buffer.from(raw, 'binary').toString('hex');
+        // Store the HEXified signature. At least useful in tests.
+        this.lastSignature = signature;
+
+        // Pad the signature with zeroes so the it is the same length as the placeholder
+        signature += Buffer
+            .from(String.fromCharCode(0).repeat((placeholderLength / 2) - raw.length))
+            .toString('hex');
+
+        // Place it in the document.
+        pdf = Buffer.concat([
+            pdf.slice(0, byteRange[1]),
+            Buffer.from(`<${signature}>`),
+            pdf.slice(byteRange[1]),
+        ]);
+
+        // Magic. Done.
+        return pdf;
+    }
+}
+
+export default new SignPdf();