+++ src/helpers/index.js
@@ -1,7 +1,0 @@
-export {default as extractSignature} from './extractSignature';
-export {default as pdfkitAddPlaceholder} from './pdfkitAddPlaceholder';
-export {default as plainAddPlaceholder} from './plainAddPlaceholder';
-export {default as removeTrailingNewLine} from './removeTrailingNewLine';
-export {default as findByteRange} from './findByteRange';
-
-'This string is added so that jest collects coverage for this file'; // eslint-disable-line