+++ src/helpers/index.test.js
@@ -1,7 +1,0 @@
-import * as helpers from './index';
-
-describe('Helpers index', () => {
-    it('Exports expected helpers', () => {
-        expect(Object.keys(helpers).sort()).toMatchSnapshot();
-    });
-});